finish

+++ db/migrate/20140907060633_add_money_to_rides.rb
@@ -1,5 +1,5 @@
 class AddMoneyToRides < ActiveRecord::Migration
   def change
-    add_column :rides, :cost, :integer
+    add_column :rides, :cost, :decimal
   end
 end
